Fix for Win x64 warning.

+++ src/maidsafe/encrypt/tests/self_encryptor_test.cc
@@ -1176,14 +1176,18 @@ TEST_F(BasicTest, BEH_DeleteStoredChunkFromDisk) {
   std::string original(RandomString(size)), recovered(size, 0);
   EXPECT_TRUE(self_encryptor_->Write(original.data(), static_cast<uint32_t>(original.size()), 0));
   EXPECT_TRUE(self_encryptor_->Flush());
-  EXPECT_TRUE(self_encryptor_->Read(const_cast<char*>(recovered.data()), recovered.size(), 0));
+  EXPECT_TRUE(self_encryptor_->Read(const_cast<char*>(recovered.data()),
+                                    static_cast<uint32_t>(recovered.size()),
+                                    0));
   fs::directory_iterator it(file_chunk_store_path_);
   while (fs::is_regular_file(it->path()))
     ++it;
   EXPECT_FALSE(it == fs::directory_iterator());
   EXPECT_GE(fs::remove_all(it->path(), error_code), 6);  // for directory depth == 5
   EXPECT_EQ(0, error_code.value());
-  EXPECT_FALSE(self_encryptor_->Read(const_cast<char*>(recovered.data()), recovered.size(), 0));
+  EXPECT_FALSE(self_encryptor_->Read(const_cast<char*>(recovered.data()),
+                                     static_cast<uint32_t>(recovered.size()),
+                                     0));
 }
 
 TEST_F(BasicTest, BEH_ManualCheckWrite) {
